Add booking resource and testing
Implement booking entity with bidirectional many-to-one association
with customer and flight entity.

Testing has been created for basic opration including create, get and
delete. A very hidden problem has been revealed and resolved, which is
creating a booking with non-existence customer/fight id or existence
customer/flight id but wrong message, now the creating API will check
both customer/flight message to make sure they are exist and correct
regrad the database.

The delete method should be tested with customer resource together, to
make sure once we delete a booking, the customer entity associated with
it would also delete the record. This should work fine but I didn't have
time to test it detailly. The test will be updated shortly to check
this.